Story #14267: Migrate ingest-external layer to ingest.
