+++ api/api-gateway/src/main/resources/application-dev.yml
@@ -276,7 +276,7 @@ spring:
             - RewritePath=/collect-api/transactions/archive-units(?<segment>.*),/collect-api/v1/transactions$\{segment},
             - RewritePath=/collect-api/transactions(?<segment>.*),/collect-api/v1/transactions$\{segment},
 
-        - id: ingest-external-service
+        - id: ingest-service
           uri: https://localhost:8088 # URL INGEST EXTERNAL
           predicates:
             - >
