+++ api/api-ingest/pom.xml
@@ -14,7 +14,7 @@
     <description>VITAMUI Ingest API</description>
 
     <modules>
-        <module>ingest-external</module>
+        <module>ingest</module>
         <module>ingest-commons</module>
     </modules>
