+++ deployment/roles/vitamui/templates/ingest/application.yml.j2
@@ -47,7 +47,7 @@ management:
 gateway:
   enabled: true
 
-ingest-external:
+ingest:
   security-client:
     server-host: {{ vitamui.security.host }}
     server-port: {{ vitamui.security.port_service }}
