+++ deployment/roles/vitamui/templates/ingest/collect-external-client.conf.j2
