+++ docs/swagger/api-external/ingest/swagger.json
