+++ intellij-conf/.run/Spring/ingest external.run.xml
@@ -1,12 +1,12 @@
 <component name="ProjectRunConfigurationManager">
-    <configuration default="false" name="ingest external" type="SpringBootApplicationConfigurationType"
+    <configuration default="false" name="ingest" type="SpringBootApplicationConfigurationType"
                    factoryName="Spring Boot">
         <option name="ACTIVE_PROFILES" value="dev"/>
         <option name="FRAME_DEACTIVATION_UPDATE_POLICY" value="UpdateClassesAndResources"/>
-        <module name="ingest-external"/>
+        <module name="ingest"/>
         <option name="SPRING_BOOT_MAIN_CLASS" value="fr.gouv.vitamui.ingest.server.ApiIngestApplication"/>
         <option name="VM_PARAMETERS" value="-Xmx100m -Dvitam.config.folder=$vitamLocalShareFolder$/conf/vitam-ui"/>
-        <option name="WORKING_DIRECTORY" value="file://$PROJECT_DIR$/api/api-ingest/ingest-external"/>
+        <option name="WORKING_DIRECTORY" value="file://$PROJECT_DIR$/api/api-ingest/ingest"/>
         <method v="2">
             <option name="Make" enabled="true"/>
         </method>
